+++ tests/common/mod.rs
@@ -7,7 +7,7 @@ use rust_mqtt::{
         Client,
         options::{ConnectOptions, DisconnectOptions, RetainHandling, SubscriptionOptions},
     },
-    config::{KeepAlive, SessionExpiryInterval},
+    config::{KeepAlive, MaximumPacketSize, SessionExpiryInterval},
     types::{MqttBinary, MqttString, QoS},
 };
 use tokio::net::TcpStream;
@@ -34,6 +34,7 @@ pub const NO_SESSION_CONNECT_OPTIONS: &ConnectOptions<'static> = &ConnectOptions
     clean_start: true,
     keep_alive: KeepAlive::Infinite,
     session_expiry_interval: SessionExpiryInterval::EndOnDisconnect,
+    maximum_packet_size: MaximumPacketSize::Unlimited,
     request_response_information: false,
     user_name: Some(USERNAME),
     password: Some(PASSWORD),